backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.7.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.3.0
python-dotenv>=1.0.1
rank-bm25>=0.2.2
pydantic-settings>=2.0.0
numpy>=1.24.0
openai>=1.0.0
